refactor(events): rename EventAnnouncementZhBiweekly -> ZhBiweeklyAnnouncement
